changed QR to appear on hero component
