+++ src/components/Hero/Hero.css
@@ -1,3 +1,15 @@
+/* body * {
+    color:                      hsla(210, 100%, 100%, 0.88)     !important;
+    background:                 hsla(210, 100%,  50%, 0.33)       !important;
+    outline:    0.25rem solid hsla(210, 100%, 100%, 0.50)     !important;
+}
+
+.debug-center {
+    display:            flex;
+    justify-content:    center;
+    align-items:        center;
+} */
+
 .hero {
     height: 500px;
     width: 100%;
@@ -7,12 +19,29 @@
     align-items: center;  
     background: linear-gradient(to top, #3204fdba, #9907facc), url("https://spacecubed.com/wp-content/uploads/2020/06/pssmileproduction_shecodes-115-scaled.jpg") no-repeat center;
     text-align: center;
+    align-items: space-between;
 }
 
 .hero-highlight {
     font-family: var(--highlight-text);
     font-size: 80px;
     text-shadow: var(--hero-highlight);
+}
+
+.qr-code-hero {
+    /* padding: 0; */
+    align-items: flex-end;
+}
+
+.qr-code {
+    /* padding: 50px; */
+    margin: 5px;
+    color: var(--primary);
+    background-color: var(--primary);
+}
+
+div.qr-code {
+    margin-top: 0;
 }
 
 /* make sure you check your index.css */
