--- src/components/QRCode/QRCode.css
@@ -1,6 +1,0 @@
-.qr-code {
-    padding: 50px;
-    margin: 500px;
-    color: var(--primary);
-    background-color: var(--primary);
-}
